1차 빌딩 성공
남은 일
Componet::RigidComponent 다시 만들기
Manager의 Create, Factory부분 다 다시 뜯어 고치기
Component, Scene의 금단의 Public 고치기

+++ 4Q_Solution/Engine/EasingFunction.h
@@ -181,7 +181,7 @@ namespace Engine::Math
 	template<typename T>
 	T EasingInterpolation(T& a, T& b, float t, EasingEffect easingEffect = EasingEffect::Linear)
 	{
-		return a + (b - a) * EasingFunction[easingEffect](t);
+		return a + (b - a) * EasingFunction[static_cast<unsigned int>(easingEffect)](t);
 	}
 }
